Bug fix: read as long, not as int

+++ src/main/java/io/github/dsheirer/module/decode/p25/phase1/message/tsbk/standard/osp/FrequencyBandUpdateVUHF.java
@@ -115,7 +115,7 @@ public class FrequencyBandUpdateVUHF extends OSPMessage implements IFrequencyBan
      */
     public boolean hasTransmitOffset()
     {
-        return getMessage().getInt(TRANSMIT_OFFSET) != 0x80;
+        return getMessage().getLong(TRANSMIT_OFFSET) != 0x80;
     }
 
     @Override
